+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - 
 ### Fixed:
 - Fixed purified bucket compatibility (by Tiviacz1337)
+- Fixed dispenser crash
 ### Changed:
 -